Add SpeechStore
