+++ Demo/DemoChat/Sources/MiscStore.swift
@@ -7,7 +7,6 @@
 
 import UIKit
 import OpenAI
-import AVFAudio
 
 public final class MiscStore: ObservableObject {
     public var openAIClient: OpenAIProtocol
@@ -92,45 +91,4 @@ public final class MiscStore: ObservableObject {
             moderationConversationError = error
         }
     }
-    
-    // MARK: - Speech
-    struct AudioObject: Identifiable {
-        let id = UUID()
-        let prompt: String
-        let audioPlayer: AVAudioPlayer?
-        let originResponse: AudioSpeechResult
-        let format: String
-    }
-    
-    @Published var audioObjects: [AudioObject] = []
-    
-    @MainActor
-    func createSpeech(_ query: AudioSpeechQuery) async {
-        guard let input = query.input, !input.isEmpty else { return }
-        do {
-            let response = try await openAIClient.audioCreateSpeech(query: query)
-            guard let data = response.audioData else { return }
-            let player = try? AVAudioPlayer(data: data)
-            let audioObject = AudioObject(prompt: input,
-                                          audioPlayer: player,
-                                          originResponse: response,
-                                          format: query.responseFormat.rawValue)
-            audioObjects.append(audioObject)
-        } catch {
-            print(error.localizedDescription)
-        }
-    }
-    
-    func getFileInDocumentsDirectory(_ data: Data, fileName: String, _ dir: @escaping (URL) -> Void) {
-        if let fileURL = try? FileManager.default.url(for: .documentDirectory, in: .userDomainMask, appropriateFor: nil, create: true) {
-            let saveURL = fileURL.appendingPathComponent(fileName)
-            do {
-                try data.write(to: saveURL)
-                dir(saveURL)
-            } catch {
-                print(error.localizedDescription)
-            }
-        }
-    }
-    
 }
